Huge refactor, fixes to collision, timing, physics

+++ src/particles.h
@@ -4,7 +4,7 @@
 #include <SDL2/SDL.h>
 
 void particles_spawn(SDL_Point pos, float xv, float yv, int amount);
-void particles_update(int delta);
+void particles_update(double delta);
 void particles_draw(void);
 
 #endif
